backend/requirements.txt
fastapi
uvicorn
google-genai
composio-core
composio-gemini
python-dotenv
pydantic
